+++ src/cone/ugm/model/groups.py
@@ -26,7 +26,7 @@ _ = TranslationStringFactory('cone.ugm')
         'groups_node_description',
         default='Container for Groups'
     ),
-    icon='ion-person-stalker',
+    icon='bi bi-people',
     addables=['group'])
 @plumbing(
     NodeInit,
@@ -38,7 +38,8 @@ class Groups(object):
     @instance_property
     def properties(self):
         props = Properties()
-        props.in_navtree = True
+        # props.in_navtree = True
+        props.skip_mainmenu = True
         return props
 
     @instance_property