+++ src/cone/ugm/settings.py
@@ -45,12 +45,12 @@ class UGMSettings(SettingsNode):
 
 @node_info(
     name='ugm_general_settings',
-    title=_('ugm_settings_node', default='UGM Settings'),
+    title=_('ugm_settings_node', default='UGM'),
     description = _(
         'ugm_settings_node_description',
         default='General user and group management settings'
     ),
-    icon='ion-person-stalker')
+    icon='bi-people')
 class GeneralSettings(UGMSettings):
     category = _('category_ugm', default='User and Group Management')
 
@@ -61,12 +61,12 @@ class GeneralSettings(UGMSettings):
 
 @node_info(
     name='ugm_localmanager_settings',
-    title=_('localmanager_settings_node', default='Local Manager Settings'),
+    title=_('localmanager_settings_node', default='Local Manager'),
     description=_(
         'localmanager_settings_node_description',
-        default='Local Manager Settings'
+        default='Manage Local Manager access rules'
     ),
-    icon='ion-person')
+    icon='bi-person-video2')
 @plumbing(Attributes)
 class LocalManagerSettings(SettingsNode):
     category = _('category_ugm', default='User and Group Management')