+++ src/cone/ugm/tests/test_browser_group.py
@@ -53,7 +53,7 @@ class BrowserGroupTests(object):
 
         with self.layer.authenticated('manager'):
             res = render_tile(group, request, 'leftcolumn')
-        expected = '<div class="column left_column col-md-6">'
+        expected = '<div class="card column left_column bg-primary-100">'
         self.assertTrue(res.find(expected) > -1)
 
     @testing.principals(
@@ -82,7 +82,7 @@ class BrowserGroupTests(object):
 
         with self.layer.authenticated('manager'):
             res = render_tile(group, request, 'rightcolumn')
-        expected = '<div class="column right_column col-md-6">'
+        expected = '<div class="card column right_column">'
         self.assertTrue(res.find(expected) > -1)
 
     @testing.principals(
@@ -116,7 +116,7 @@ class BrowserGroupTests(object):
         with self.layer.authenticated('manager'):
             res = render_tile(group, request, 'columnlisting')
         expected = (
-            '<li class="list-group-item "\n              '
+            '<li class="list-group-item d-flex justify-content-between "\n              '
             'ajax:target="http://example.com/users/user_1">'
         )
         self.assertTrue(res.find(expected) > -1)
@@ -155,13 +155,13 @@ class BrowserGroupTests(object):
         with self.layer.authenticated('manager'):
             res = render_tile(group, request, 'allcolumnlisting')
         expected = (
-            '<li class="list-group-item "\n              '
+            '<li class="list-group-item d-flex justify-content-between "\n              '
             'ajax:target="http://example.com/users/user_1">'
         )
         self.assertTrue(res.find(expected) > -1)
 
         expected = (
-            '<li class="list-group-item "\n              '
+            '<li class="list-group-item d-flex justify-content-between "\n              '
             'ajax:target="http://example.com/users/user_2">'
         )
         self.assertTrue(res.find(expected) > -1)